open_street_maps/requirements.txt
Django==1.11.17
django-osm-field==0.3.1
pytz==2019.3
six==1.12.0
